Refactor utilities and types to improve consistency and maintainability
- Replaced toast notifications with a custom notify function in UtilitiesTextPage and UtilitiesTimerPage for better error handling.
- Removed unused memoized values in UtilitiesTimerPage to simplify state management.
- Updated audio-store and presentation-store to import types from bindings for better organization.
- Introduced a new queue-store for managing audio playback queue.
- Consolidated various type definitions across multiple files to import from a centralized bindings file.
- Cleaned up legacy fetch types and removed unused interfaces in favor of bindings.
- Updated tests to reflect changes in monitor configuration properties for consistency.

+++ src-tauri/src/db/models.rs
@@ -1,7 +1,10 @@
 use serde::{Deserialize, Serialize};
+use specta::Type;
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct Hymn {
+    #[specta(type = i32)]
     pub id: i64,
     pub number: Option<i64>,
     pub title: String,
@@ -15,12 +18,14 @@ pub struct Hymn {
     pub notes: Option<String>,
     pub cover_path: Option<String>,
     pub lyrics_sync: Option<String>,
+    #[specta(type = i32)]
     pub api_music_id: Option<i64>,
     pub created_at: String,
     pub updated_at: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct HymnWriteInput {
     pub number: Option<i64>,
     pub title: String,
@@ -36,15 +41,17 @@ pub struct HymnWriteInput {
     pub lyrics_sync: Option<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct Album {
     pub name: String,
-    pub hymn_count: i64,
+    pub hymn_count: i32,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct BibleVersion {
+    #[specta(type = i32)]
     pub id: i64,
     pub name: String,
     pub abbreviation: String,
@@ -52,25 +59,27 @@ pub struct BibleVersion {
     pub file_path: Option<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct Book {
     pub name: String,
-    pub chapter_count: i64,
+    pub chapter_count: i32,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct Verse {
+    #[specta(type = i32)]
     pub id: i64,
+    #[specta(type = i32)]
     pub version_id: i64,
     pub book: String,
-    pub chapter: i64,
-    pub verse: i64,
+    pub chapter: i32,
+    pub verse: i32,
     pub text: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct BibleSearchResult {
     pub verse: Verse,
@@ -78,8 +87,10 @@ pub struct BibleSearchResult {
     pub snippet: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct Presentation {
+    #[specta(type = i32)]
     pub id: i64,
     pub title: String,
     pub author: Option<String>,
@@ -90,20 +101,24 @@ pub struct Presentation {
     pub updated_at: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct Slide {
+    #[specta(type = i32)]
     pub id: i64,
+    #[specta(type = i32)]
     pub presentation_id: i64,
-    pub slide_index: i64,
+    pub slide_index: i32,
     pub slide_type: String,
     pub content: String,
     pub notes: Option<String>,
     pub transition: Option<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct Service {
+    #[specta(type = i32)]
     pub id: i64,
     pub title: String,
     pub date: Option<String>,
@@ -112,40 +127,47 @@ pub struct Service {
     pub updated_at: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct ServiceItem {
+    #[specta(type = i32)]
     pub id: i64,
+    #[specta(type = i32)]
     pub service_id: i64,
     pub item_type: String,
+    #[specta(type = i32)]
     pub item_id: Option<i64>,
     pub title: String,
-    pub item_order: i64,
+    pub item_order: i32,
     pub notes: Option<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct ServiceWithItems {
     pub service: Service,
     pub items: Vec<ServiceItem>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct Setting {
     pub key: String,
     pub value: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct MonitorConfig {
+    #[specta(type = i32)]
     pub id: i64,
     pub monitor_id: String,
     pub role: String,
     pub enabled: bool,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct MonitorInfo {
     pub id: String,
     pub name: String,
@@ -161,43 +183,34 @@ pub struct MonitorInfo {
     pub scale_factor: f64,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct SlideContent {
     pub slide_type: String,
     pub text: Option<String>,
     pub title: Option<String>,
     pub subtitle: Option<String>,
     pub label: Option<String>,
-    #[serde(default, alias = "videoPath")]
     pub video_path: Option<String>,
-    #[serde(default, alias = "backgroundImage")]
     pub background_image: Option<String>,
-    #[serde(default, alias = "backgroundColor")]
     pub background_color: Option<String>,
-    #[serde(default, alias = "audioPath")]
     pub audio_path: Option<String>,
-    #[serde(default, alias = "autoPlay")]
     pub auto_play: Option<bool>,
-    #[serde(default, alias = "loop")]
     pub r#loop: Option<bool>,
-    #[serde(default)]
     pub muted: Option<bool>,
-    #[serde(default)]
     pub mode: Option<String>,
-    #[serde(default, alias = "textColor")]
     pub text_color: Option<String>,
-    #[serde(default, alias = "textSize")]
     pub text_size: Option<i32>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct OverlayState {
     pub black_screen: bool,
     pub logo_screen: bool,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct SlideContext {
     pub next: Option<SlideContent>,
@@ -206,33 +219,38 @@ pub struct SlideContext {
     pub title: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct VideoMetadata {
+    #[specta(type = f64)]
     pub duration_ms: i64,
     pub width: i32,
     pub height: i32,
+    #[specta(type = f64)]
     pub file_size: i64,
     pub format: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct Collection {
+    #[specta(type = i32)]
     pub id: i64,
     pub name: String,
     pub description: Option<String>,
     pub year: Option<i32>,
     pub cover_path: Option<String>,
     pub auto_cover_path: Option<String>,
-    pub song_count: i64,
+    pub song_count: i32,
     pub source_type: String,
+    #[specta(type = i32)]
     pub api_album_id: Option<i64>,
     pub created_at: String,
     pub updated_at: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub enum CollectionSongSyncStatus {
     InSync,
     Stale,
@@ -261,33 +279,42 @@ impl CollectionSongSyncStatus {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct CollectionSong {
+    #[specta(type = i32)]
     pub id: i64,
+    #[specta(type = i32)]
     pub collection_id: i64,
     pub source_path: String,
     pub source_format: String,
     pub source_hash: Option<String>,
+    #[specta(type = i32)]
     pub source_mtime_ms: Option<i64>,
+    #[specta(type = i32)]
     pub cache_presentation_id: Option<i64>,
     pub sync_status: CollectionSongSyncStatus,
     pub last_sync_at: Option<String>,
-    pub item_order: i64,
+    pub item_order: i32,
     pub created_at: String,
     pub updated_at: String,
     pub cache_presentation_title: Option<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct CollectionWithSongs {
     pub collection: Collection,
     pub songs: Vec<CollectionSong>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
+#[serde(rename_all = "camelCase")]
 pub struct CollectionSearchResult {
     pub kind: String,
+    #[specta(type = i32)]
     pub collection_id: i64,
+    #[specta(type = i32)]
     pub song_id: Option<i64>,
     pub collection_name: String,
     pub title: String,
@@ -297,20 +324,23 @@ pub struct CollectionSearchResult {
 /// Join table struct for collection ↔ hymn links.
 /// Scaffolded for future direct-link CRUD endpoints; current queries return `Vec<Hymn>` instead.
 #[allow(dead_code)]
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct CollectionHymn {
+    #[specta(type = i32)]
     pub id: i64,
+    #[specta(type = i32)]
     pub collection_id: i64,
+    #[specta(type = i32)]
     pub hymn_id: i64,
-    pub item_order: i64,
+    pub item_order: i32,
     pub created_at: String,
 }
 
 /// Collection with its linked hymns.
 /// Used by `get_collection_with_hymns` query (scaffolded for future collection detail endpoint).
 #[allow(dead_code)]
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct CollectionWithHymns {
     pub collection: Collection,
